Use mamba (#36)
* Use mamba

conda has a lurking NP-hard SAT-solver lurking inside of it
which almost at random, due to one version change or missing build
of some dependency package, can suddenly become extremely slow.

mamba has a smarter solver that takes a fraction of the time.

See https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/manage-channels.html
See https://github.com/conda/conda/issues/8051

Fixes https://github.com/shimming-toolbox/shimming-toolbox/issues/369#issuecomment-1083589939

* Correct the shebang

* Convert to ints when using Rescale

* Update fsleyes version to 1.4.6

* Bump ST version

* Use python 3.9

* Update testing procedure for M1 macs

* Fix overlay bug

* Bump version to ST to master

+++ fsleyes_plugin_shimming_toolbox/st_plugin.py
@@ -55,7 +55,7 @@ play_icon = wx.Bitmap(os.path.join(DIR, 'img', 'play.png'), wx.BITMAP_TYPE_PNG)
 rtd_logo = wx.Bitmap(os.path.join(DIR, 'img', 'RTD.png'), wx.BITMAP_TYPE_PNG)
 # Load ShimmingToolbox logo saved as a png image, rescale it, and return it as a wx.Bitmap image.
 st_logo = wx.Image(os.path.join(DIR, 'img', 'shimming_toolbox_logo.png'), wx.BITMAP_TYPE_PNG)
-st_logo.Rescale(st_logo.GetWidth() * 0.2, st_logo.GetHeight() * 0.2, wx.IMAGE_QUALITY_HIGH)
+st_logo.Rescale(int(st_logo.GetWidth() * 0.2), int(st_logo.GetHeight() * 0.2), wx.IMAGE_QUALITY_HIGH)
 st_logo = st_logo.ConvertToBitmap()
